third_party/WebKit/LayoutTests/inspector/elements/styles/styles-computed-animation-trace.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<style>
body {
  font-family: Arial;
}

div {
  width: 300px;
  height: 300px;
  background-color: red;
  -webkit-animation-name: slide;
  -webkit-animation-duration: 100s;
  -webkit-animation-iteration-count: infinite;
  -webkit-animation-play-state: paused;
  -webkit-animation-delay: -10s;
  -webkit-animation-timing-function: linear;
}

#box2 {
  top: 20px;
  left: 20px;
}

@keyframes slide {
  from {
    top: 0px;
    left: 0px;
  }

  to {
    top: 100px;
    left: 100px;
  }
}

</style>
<script>

function test()
{
    InspectorTest.selectNodeAndWaitForStylesWithComputed("box1", step1);

    function step1()
    {
        InspectorTest.addResult("==== Computed style for box1 ====");
        InspectorTest.dumpSelectedElementStyles(false, true);
        InspectorTest.selectNodeAndWaitForStylesWithComputed("box2", step2);
    }

    function step2()
    {
        InspectorTest.addResult("==== Computed style for box2 ====");
        InspectorTest.dumpSelectedElementStyles(false, true);
        InspectorTest.completeTest();
    }
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that computed styles pane shows animation applied styles.
</p>

<div id="box1"></div>
<div id="box2"></div>

</body>
</html>
