third_party/WebKit/LayoutTests/inspector/elements/styles/styles-computed-animation-trace-expected.txt

Tests that computed styles pane shows animation applied styles.

==== Computed style for box1 ====
[expanded] 
-webkit-animation-delay: -10s;
    div - -10s styles-computed…-trace.html:10
-webkit-animation-duration: 100s;
    div - 100s styles-computed…-trace.html:10
-webkit-animation-iteration-count: infinite;
    div - infinite styles-computed…-trace.html:10
-webkit-animation-name: slide;
    div - slide styles-computed…-trace.html:10
-webkit-animation-play-state: paused;
    div - paused styles-computed…-trace.html:10
-webkit-animation-timing-function: linear;
    div - linear styles-computed…-trace.html:10
animation-delay: -10s;
    div - -10s styles-computed…-trace.html:10
animation-duration: 100s;
    div - 100s styles-computed…-trace.html:10
animation-iteration-count: infinite;
    div - infinite styles-computed…-trace.html:10
animation-name: slide;
    div - slide styles-computed…-trace.html:10
animation-play-state: paused;
    div - paused styles-computed…-trace.html:10
animation-timing-function: linear;
    div - linear styles-computed…-trace.html:10
background-color: rgb(255, 0, 0);
    div - red styles-computed…-trace.html:10
display: block;
    div - block user agent stylesheet
font-family: Arial;
    body - Arial styles-computed…n-trace.html:6
left: 10px;
    Animation slide
top: 10px;
    Animation slide

[expanded] 
element.style  { ()


==== Computed style for box2 ====
[expanded] 
-webkit-animation-delay: -10s;
    div - -10s styles-computed…-trace.html:10
-webkit-animation-duration: 100s;
    div - 100s styles-computed…-trace.html:10
-webkit-animation-iteration-count: infinite;
    div - infinite styles-computed…-trace.html:10
-webkit-animation-name: slide;
    div - slide styles-computed…-trace.html:10
-webkit-animation-play-state: paused;
    div - paused styles-computed…-trace.html:10
-webkit-animation-timing-function: linear;
    div - linear styles-computed…-trace.html:10
animation-delay: -10s;
    div - -10s styles-computed…-trace.html:10
animation-duration: 100s;
    div - 100s styles-computed…-trace.html:10
animation-iteration-count: infinite;
    div - infinite styles-computed…-trace.html:10
animation-name: slide;
    div - slide styles-computed…-trace.html:10
animation-play-state: paused;
    div - paused styles-computed…-trace.html:10
animation-timing-function: linear;
    div - linear styles-computed…-trace.html:10
background-color: rgb(255, 0, 0);
    div - red styles-computed…-trace.html:10
display: block;
    div - block user agent stylesheet
font-family: Arial;
    body - Arial styles-computed…n-trace.html:6
left: 10px;
    Animation slide
    #box2 - 20px styles-computed…-trace.html:22
top: 10px;
    Animation slide
    #box2 - 20px styles-computed…-trace.html:22

[expanded] 
element.style  { ()
